requirements.txt
fastapi
uvicorn
transformers
accelerate
sentencepiece
torch
sentence-transformers
faiss-cpu
pypdf
python-multipart
numpy
regex
